@@ -8,23 +8,24 @@ authors = [{name = "Tri Songz", email = "ts@songzcorp.com"}]
 classifiers = [
     "Intended Audience :: Developers",
     "License :: OSI Approved :: MIT License",
-    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
     "Topic :: Software Development :: Libraries",
 ]
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 dependencies = [
     "loguru",
-    "pydantic",
-    "pydantic-settings",
+    "pydantic>=2.0.0",
+    "pydantic-settings>=2.0.0",
     "frozendict",
     "async_lru",
     "pyyaml",
     "filelock",
     "xxhash",
-    "typing_extensions; python_version < '3.8'",
-    "setuptools; python_version > '3.11'",
     "anyio>=3.7.1",
     "typer>=0.17.3",
+    "httpx>=0.28.1",
 ]
 
 [urls]
@@ -34,6 +35,8 @@ Homepage = "https://github.com/trisongz/lazyops"
 requires = ["setuptools"]
 build-backend = "setuptools.build_meta"
 
+[tool.uv.sources]
+kvdb-py = { git = "https://github.com/trisongz/kvdb-py" }
 
 [project.scripts]
 lzl = "lzl.cmd:main"
@@ -63,13 +66,12 @@ openai = [
     "numpy",
     "scipy",
 ]
-# Maintain compatability for now.
 fileio = [
-    "botocore>=1.34.0",
-    "boto3>=1.34.0",
-    "aiobotocore>=2.11.0",
-    "fsspec>=2024.2.0",
-    "s3fs>=2024.2.0",
+    "botocore>=1.35.0",
+    "boto3>=1.35.0",
+    "aiobotocore>=2.15.0",
+    "fsspec>=2024.10.0",
+    "s3fs>=2024.10.0",
     "s3transfer>=0.10.0",
     "python-magic",
     "aiopath",
@@ -77,11 +79,11 @@ fileio = [
     "aiofile",
 ]
 file = [
-    "botocore>=1.34.0",
-    "boto3>=1.34.0",
-    "aiobotocore>=2.11.0",
-    "fsspec>=2024.2.0",
-    "s3fs>=2024.2.0",
+    "botocore>=1.35.0",
+    "boto3>=1.35.0",
+    "aiobotocore>=2.15.0",
+    "fsspec>=2024.10.0",
+    "s3fs>=2024.10.0",
     "s3transfer>=0.10.0",
     "python-magic",
     "aiopath",
@@ -93,6 +95,10 @@ docs = [
     "mkdocs-material>=9.0.0",
     "mkdocstrings[python]>=0.24.0",
     "pymdown-extensions>=10.0",
+]
+e2e = [
+    "lazyops[dev,file,fastapi,kops,openai]",
+    "kvdb-py",
 ]
 
 [tool.setuptools]
@@ -114,4 +120,9 @@ builder-ci = [
     "pyyaml",
     "typing_extensions; python_version < '3.8'",
     "setuptools; python_version > '3.11'",
+]
+dev = [
+    "pytest",
+    "pytest-asyncio",
+    "pytest-cov"
 ]
