@@ -11,7 +11,7 @@ if t.TYPE_CHECKING:
     import s3fs
     import boto3
     import boto3.session
-    from ..compat.r2 import R2FileSystem
+    from ..compat.r2.filesys import R2FileSystem
     from s3transfer.manager import TransferManager
     from ...configs.providers import (
         AWSConfig,
@@ -20,6 +20,87 @@ if t.TYPE_CHECKING:
         CloudflareR2Config,
     )
 
+
+def _patch_s3_mkdir(cls):
+    _orig_mkdir = cls.mkdir
+    _orig_amkdir = cls.amkdir
+    
+    def _get_bucket(path: str) -> str:
+        # returns the first part of the path
+        return path.split('/', 1)[0]
+    
+    def mkdir(path: str, create_parents: bool = True, exist_ok: bool = True, **kwargs):
+        try:
+            return _orig_mkdir(path, create_parents=create_parents, exist_ok=exist_ok, **kwargs)
+        except (OSError, Exception) as e:
+            # s3fs raises OSError or others
+            msg = str(e)
+            if "NoSuchBucket" in msg or "The specified location-constraint is not valid" in msg or "Bucket does not exist" in msg:
+                bucket = _get_bucket(path)
+                created = False
+                try: 
+                    # Attempt to create bucket using s3fs
+                    # s3fs.mkdir(bucket) creates the bucket.
+                    cls.CloudFileSystem.fs.mkdir(bucket)
+                    created = True
+                except Exception: 
+                    pass
+                
+                if not created and cls.CloudFileSystem.boto:
+                    # Fallback to boto
+                    try:
+                        # We might need LocationConstraint
+                        config = cls.CloudFileSystem.fsconfig
+                        params = {'Bucket': bucket}
+                        region = getattr(config, 'region', None) if config else None
+                        
+                        if region and region != 'us-east-1':
+                            params['CreateBucketConfiguration'] = {'LocationConstraint': region}
+                        
+                        cls.CloudFileSystem.boto.create_bucket(**params)
+                        created = True
+                    except Exception:
+                        pass
+                
+                # Retry original operation
+                return _orig_mkdir(path, create_parents=create_parents, exist_ok=exist_ok, **kwargs)
+            raise
+
+    async def amkdir(path: str, create_parents: bool = True, exist_ok: bool = True, **kwargs):
+        try:
+            return await _orig_amkdir(path, create_parents=create_parents, exist_ok=exist_ok, **kwargs)
+        except (OSError, Exception) as e:
+            msg = str(e)
+            if "NoSuchBucket" in msg or "The specified location-constraint is not valid" in msg or "Bucket does not exist" in msg:
+                bucket = _get_bucket(path)
+                try: 
+                    # Try to create bucket using aiobotocore client directly
+                    fsa = cls.CloudFileSystem.fsa
+                    
+                    if hasattr(fsa, 's3') and fsa.s3:
+                         s3_client = fsa.s3
+                         if s3_client is not None:
+                            config = cls.CloudFileSystem.fsconfig
+                            params = {'Bucket': bucket}
+                            region = getattr(config, 'region', None) if config else None
+
+                            if region and region != 'us-east-1':
+                                params['CreateBucketConfiguration'] = {'LocationConstraint': region}
+                            
+                            await s3_client.create_bucket(**params)
+
+                    elif hasattr(fsa, '_mkdir'):
+                        # Fallback to internal mkdir if s3 client not exposed (unlikely)
+                        await fsa._mkdir(bucket)
+                    
+                except Exception: 
+                    pass
+                # Retry
+                return await _orig_amkdir(path, create_parents=create_parents, exist_ok=exist_ok, **kwargs)
+            raise
+
+    cls.mkdir = mkdir
+    cls.amkdir = amkdir
 
 class AWSFileSystem(metaclass = CloudFileSystemMeta):
     fs: 's3fs.S3FileSystem' = None
@@ -37,6 +118,11 @@ class AWSAccessor(BaseFileSystemAccessor):
 
     class CloudFileSystem(AWSFileSystem):
         pass
+
+    @classmethod
+    def reload_cfs(cls, **kwargs):
+        super().reload_cfs(**kwargs)
+        _patch_s3_mkdir(cls)
 
 
 class MinioFileSystem(metaclass = CloudFileSystemMeta):
@@ -57,6 +143,11 @@ class MinioAccessor(BaseFileSystemAccessor):
     class CloudFileSystem(MinioFileSystem):
         pass
 
+    @classmethod
+    def reload_cfs(cls, **kwargs):
+        super().reload_cfs(**kwargs)
+        _patch_s3_mkdir(cls)
+
 
 class S3CFileSystem(metaclass = CloudFileSystemMeta):
     fs: 's3fs.S3FileSystem' = None
@@ -75,6 +166,11 @@ class S3CAccessor(BaseFileSystemAccessor):
     class CloudFileSystem(S3CFileSystem):
         pass
 
+    @classmethod
+    def reload_cfs(cls, **kwargs):
+        super().reload_cfs(**kwargs)
+        _patch_s3_mkdir(cls)
+
 class R2FileSystem(metaclass = CloudFileSystemMeta):
     fs: 'R2FileSystem' = None
     fsa: 'R2FileSystem' = None
@@ -89,4 +185,10 @@ class R2Accessor(BaseFileSystemAccessor):
 
     class CloudFileSystem(R2FileSystem):
         pass
+    
+    # R2 usually needs explicit bucket creation too? 
+    # But sticking to S3 ones for now.
+    # If R2 uses s3fs-like, it might also benefit.
+    # Leaving standard.
+
 
